+++ scrypto-compiler/tests/compilation.rs
@@ -1,8 +1,11 @@
 #[cfg(test)]
 mod tests {
     use radix_common::prelude::*;
+    use radix_engine::utils::ExtractSchemaError;
+    use radix_engine::vm::wasm::PrepareError;
     use radix_engine_interface::types::Level;
     use scrypto_compiler::*;
+    use std::process::Command;
     use std::{env, path::PathBuf, process::Stdio};
     use tempdir::TempDir;
 
@@ -452,5 +455,76 @@ mod tests {
 
         // Assert
         assert!(status.is_ok(), "{:?}", status);
+    }
+
+    #[test]
+    fn test_compilation_with_wasm_reference_types_disabled() {
+        // Arrange
+        let mut blueprint_manifest_path = PathBuf::from(env!("CARGO_MANIFEST_DIR"));
+
+        blueprint_manifest_path.extend(["tests", "assets", "call_indirect", "Cargo.toml"]);
+
+        // Act
+        // ScryptoCompiler compiles WASM by default with reference-types disabled.
+        let status = ScryptoCompiler::builder()
+            .manifest_path(blueprint_manifest_path)
+            .compile();
+
+        // Assert
+        assert!(status.is_ok(), "{:?}", status);
+    }
+
+    #[test]
+    fn test_compilation_with_wasm_reference_types_enabled() {
+        // Arrange
+        let mut blueprint_manifest_path = PathBuf::from(env!("CARGO_MANIFEST_DIR"));
+
+        blueprint_manifest_path.extend(["tests", "assets", "call_indirect", "Cargo.toml"]);
+
+        // Check clang/LLVM version
+        let clang_version = Command::new("clang").arg("--version").output().unwrap();
+
+        // clang --version exemplary output
+        // Ubuntu clang version 17.0.6 (++20231209124227+6009708b4367-1~exp1~20231209124336.77)
+        // Target: x86_64-pc-linux-gnu
+        // Thread model: posix
+        // InstalledDir: /usr/lib/llvm-17/bin
+        let clang_version = String::from_utf8_lossy(&clang_version.stdout);
+        let mut idx = clang_version
+            .find("clang version")
+            .expect("Failed to get clang version");
+        idx += "clang version ".len();
+        let version = &clang_version[idx..]
+            .split_whitespace()
+            .next()
+            .expect("Failed to get version");
+        let major_version = version
+            .split(".")
+            .next()
+            .expect("Failed to get major version");
+        let major_version: u8 = major_version.parse().unwrap();
+
+        let action = if major_version >= 19 {
+            // Since LLVM 19 reference-types are enabled by default, no dedicated CFLAGS needed.
+            // Unset TARGET_CFLAGS to build with default WASM features.
+            EnvironmentVariableAction::Unset
+        } else {
+            // In previous versions reference-types must be enabled explicitly.
+            EnvironmentVariableAction::Set("-mreference-types".to_string())
+        };
+        // Act
+        let status = ScryptoCompiler::builder()
+            .env("TARGET_CFLAGS", action)
+            .manifest_path(blueprint_manifest_path)
+            .compile();
+
+        // Assert
+        // Error is expected here because Radix Engine expects WASM with reference-types disabled.
+        // See `call_indirect.c` for more details.
+        assert!(matches!(
+            status.unwrap_err(),
+            ScryptoCompilerError::SchemaExtractionError(
+                ExtractSchemaError::InvalidWasm(PrepareError::ValidationError(msg))) if msg.contains("reference-types not enabled: zero byte expected")
+        ))
     }
 }
